Moved Swarm under swarm/Swarm.ts
