+++ src/components/Card.tsx
@@ -7,7 +7,7 @@ import { TouchableView } from './TouchableView';
 import { DateUtils } from '../DateUtils';
 import { Utils } from '../Utils';
 import { ImageView } from './ImageView';
-import { isSwarmLink } from '../Swarm';
+import { isSwarmLink } from '../swarm/Swarm';
 import { ImageData } from '../models/ImageData';
 import Markdown from 'react-native-easy-markdown';
 import { ErrorBoundary } from './ErrorBoundary';
