+++ src/components/FeedInfo.tsx
@@ -17,7 +17,7 @@ import { Feed } from '../models/Feed';
 import { SimpleTextInput } from './SimpleTextInput';
 import { Debug } from '../Debug';
 import { Colors } from '../styles';
-import * as Swarm from '../Swarm';
+import * as Swarm from '../swarm/Swarm';
 import { downloadPostFeed } from '../PostFeed';
 import { NavigationHeader } from './NavigationHeader';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
